Implement recursive field uncovering and game loss condition
Added recursive field uncovering logic to the UncoverFieldCommand. This enables fields without any mines nearby to be uncovered automatically, improving the game play experience. The game loss condition when a mine is uncovered was also implemented. In addition, added functions to SquareGridMinefieldTopology to support calculating nearby fields and mines more effectively, as well as tests for these.